add repoParent and repoSource fields
These are populated with the repo that it was forked from, and the top repo
in the graph, respectively. This works when using userRepo to get a Repo.

Note that repoParent and repoSource are RepoRef and not Repo because the
API only returns an abbreviated object for these, not sufficient to
construct a full repo.

Closes https://github.com/fpco/github/issues/43