Fix data declarations for Content results
Previously, the declarations made the assumption that every
`ContentData` has an encoding, size, and data. However, only files (not
directories) have this information. Additionally, you only get this
additional file information when you specifically request a file; not
when a file is in a directory listing from getting the contents of a
directory.

Having read the GitHub api docs, it seems there are three types of
things we want to differentiate:

* When you get the contents of a file - those contents
* When you get the contents of a directory - a file which is in that
  directory
* When you get the contents of a directory - a subdirectory

This commit fixes the data declarations to align better with the API.
It adds a few new types:

* ContentInfo, which is common to all three of the above things
* ContentFileData, the first one
* ContentItem, which covers the second and third
* ContentItemType, which allows you to tell the difference between
  the second and third

It also removes ContentData, which is subsumed by these new types.

I've tested all three cases in the REPL (on a public repo only) and
this appears to work. Updates to samples are soon to come.