Resolve warning regarding parseUrlThrow
The http-client library introduced a new name for `parseUrl` in version 0.4.30,
`parseUrlThrow`, and deprecated the use of `parseUrl`. This causes a warning
during compilation. Since we support http-client above `0.4.8.1`, I have added
some additional CPP conditional compilation to import and use `parseUrl` for
older versions of http-client, and to import and use `parseUrlThrow` in the
newer versions.