Add eventId field to Event data type.
The id for an event is a string. This means we cannot convert it to an
`Id Event`, since the `Id` data constructor can only be applied to a
value of type `Int`.

We also avoid converting the id of type String to an Int, since that is
frought with multiple ways to fail, particularly if the string
represents an Int greater than 2^29 - 1.