Fix permission field in types for teams
Add privacy field to types for teams

The privacy field was previously commented out while the privacy level
of a team was part of the `ironman` API preview. This feature is now
part of the official v3 API, so this commit adds support for this field.

Make privacy field non-optional in teams

The `privacy` field is optional when creating/editing teams, but
has a context dependent default value. When getting the team details,
the field is expected to always be populated due to these default
values.

Update team privacy/permission optionality

Makes team `privacy` and `permission` fields required when creating a
team so users are forced to choose, but optional when editing a team.

Filter null values from types for teams